Added integration with database to Sign Up
I added code to integrate with the database for the Sign Up page so that it actually successfully creates a user.  I also added some validation checks to each input field to make sure the correct data is stored in the database.  I am unsure about adding the username to the database, though.