Added some search functionality
I added basic search functionality to the Game Library such that a user is able to enter a search query and results will be returned according to the search query.  I also extracted the sort and filter widgets to a separate method for better code.  I also changed the background image to show a placeholder when no image for a game is found.