Updated Search
I updated the search so that it now only loads the searched for games and removes the infinite scrolling when a term is searched for.  As soon as the searched is cleared the normal games are loaded again as usual from the start.  I also modularised the code by moving the API request to a separate function so that many types of API requests can be made without copying code.