Added new models and fixed filter
I added new filter models (classes) to make dynamic loading of the filter page possible.  This now allows the application to fetch filter values from the RAWG API when a FilterList model is created, which is used to populate the checkbox widgets with filter data.  This also now uses a singleton pattern so that the api request is only called once.