Updated how the friend search works
I changed the search in friends to be more usable.  I removed the OnChanged() callback so that the page does not reload everytime a user changes the value of the textbox.  I also added a more visible icon button that can be clicked on to search for the entered user.